requirements.txt
flask
gunicorn
python-dotenv
psycopg2-binary
Flask-Caching
apscheduler
requests
PyJWT